refactor(profiles): technician read+write in service, thin route (debt #5 phase 2)
Gives the technician concept a single read+write domain layer. The self-service
endpoint /api/user/technician-profile was doing inline DB work + business rules
(SoC violation); now it's a thin controller.

- technician-service gains getTechnicianSelfProfile(userId) (the self-edit read
  shape) and upsertTechnicianProfile(userId, input) (upsert repairer_profiles +
  replace user_skills + the professional-demotion guard + getCategoryForSkill).
- /api/user/technician-profile GET/PUT now just validate + delegate. ~160 lines
  of inline query/mapping/business logic removed from the route.

Behavior-preserving (same upsert/skills/tier-guard semantics). typecheck + lint
clean; route + service tests pass; full suite zero new failures vs baseline.

Remaining for #5 phase 2 (follow-up): route the live /api/admin/it-hilfe/helpers
through the service (+ rename to /api/admin/technicians), consolidate the
/api/technicians list mapper, resolve the serviceTypes<->serviceDeliveryTypes alias.
