+++ src/lib/services/technician-service.ts
@@ -9,7 +9,9 @@ import { db } from '@/db'
 import { repairerProfiles, repairerServices, userSkills, users, userProfiles } from '@/db/schema'
 import { eq, and, sql } from 'drizzle-orm'
 import { logger } from '@/lib/logger'
-import { REPAIRER_PROFILE_TIER } from '@/config/repairer-status'
+import { REPAIRER_PROFILE_TIER, REPAIRER_STATUS } from '@/config/repairer-status'
+import { IT_SKILLS } from '@/config/it-hilfe'
+import type { TechnicianProfileInput } from '@/lib/schemas'
 
 export const TECHNICIAN_UUID_RE =
   /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i
@@ -178,4 +180,167 @@ export async function getTechnicianByIdOrUserId(id: string): Promise<TechnicianD
 
   if (!profile) return null
   return getTechnicianById(profile.profileId)
+}
+
+// ============================================================================
+// Self-service profile (community technicians) — read + write
+// ============================================================================
+
+export interface TechnicianSelfProfile {
+  skills: string[]
+  bio: string
+  hourlyRateCents: number | null
+  acceptsGratis: boolean
+  acceptsKulturlegi: boolean
+  serviceTypes: string[]
+  postalCode: string
+  city: string
+  canton: string
+  maxTravelKm: number
+  isActive: boolean
+  profileTier: string | null
+}
+
+/**
+ * Read the current user's own technician profile + skills (the self-service
+ * edit shape — keyed by user_id, not profile id). Returns the profile (or a
+ * skills-only stub if the user has skills but no profile row yet) and whether a
+ * profile row exists.
+ */
+export async function getTechnicianSelfProfile(
+  userId: string,
+): Promise<{ profile: TechnicianSelfProfile | null; hasProfile: boolean }> {
+  const [profileRow] = await db
+    .select({
+      bio: repairerProfiles.description,
+      hourlyRateCents: repairerProfiles.hourlyRateCents,
+      acceptsGratis: repairerProfiles.acceptsGratis,
+      acceptsKulturlegi: repairerProfiles.acceptsKulturlegi,
+      serviceTypes: repairerProfiles.serviceDeliveryTypes,
+      postalCode: repairerProfiles.postalCode,
+      city: repairerProfiles.city,
+      canton: repairerProfiles.canton,
+      maxTravelKm: repairerProfiles.maxTravelKm,
+      isActive: repairerProfiles.isActive,
+      profileTier: repairerProfiles.profileTier,
+    })
+    .from(repairerProfiles)
+    .where(eq(repairerProfiles.userId, userId))
+
+  const skillRows = await db
+    .select({ skillId: userSkills.skillId })
+    .from(userSkills)
+    .where(eq(userSkills.userId, userId))
+  const skills = skillRows.map((r) => r.skillId)
+
+  let profile: TechnicianSelfProfile | null = null
+  if (profileRow) {
+    profile = {
+      skills,
+      bio: profileRow.bio || '',
+      hourlyRateCents: profileRow.hourlyRateCents,
+      acceptsGratis: profileRow.acceptsGratis ?? true,
+      acceptsKulturlegi: profileRow.acceptsKulturlegi ?? true,
+      serviceTypes: profileRow.serviceTypes || ['flexible'],
+      postalCode: profileRow.postalCode || '',
+      city: profileRow.city || '',
+      canton: profileRow.canton || '',
+      maxTravelKm: profileRow.maxTravelKm ?? 10,
+      isActive: profileRow.isActive ?? false,
+      profileTier: profileRow.profileTier,
+    }
+  } else if (skills.length > 0) {
+    profile = {
+      skills,
+      bio: '',
+      hourlyRateCents: null,
+      acceptsGratis: true,
+      acceptsKulturlegi: true,
+      serviceTypes: ['flexible'],
+      postalCode: '',
+      city: '',
+      canton: '',
+      maxTravelKm: 10,
+      isActive: false,
+      profileTier: REPAIRER_PROFILE_TIER.COMMUNITY,
+    }
+  }
+
+  return { profile, hasProfile: !!profileRow }
+}
+
+/** Resolve a skill id to its IT_SKILLS category (for user_skills.category_id). */
+function getCategoryForSkill(skillId: string): string {
+  for (const [categoryId, skills] of Object.entries(IT_SKILLS)) {
+    if ((skills as Array<{ id: string }>).some((s) => s.id === skillId)) {
+      return categoryId
+    }
+  }
+  return 'other'
+}
+
+/**
+ * Create or update the current user's self-service (community) technician
+ * profile, then replace their skill set. Upserts repairer_profiles with
+ * tier=community but NEVER demotes an existing professional (tier guard in the
+ * conflict update). NOT-NULL placeholder columns (phone/address) default to ''.
+ */
+export async function upsertTechnicianProfile(
+  userId: string,
+  input: TechnicianProfileInput,
+): Promise<void> {
+  const {
+    skills, bio, hourlyRateCents, acceptsGratis, acceptsKulturlegi,
+    serviceTypes, postalCode, city, canton, maxTravelKm, isActive,
+  } = input
+
+  await db
+    .insert(repairerProfiles)
+    .values({
+      userId,
+      description: bio || undefined,
+      hourlyRateCents,
+      acceptsGratis,
+      acceptsKulturlegi,
+      serviceDeliveryTypes: serviceTypes.length > 0 ? serviceTypes : undefined,
+      city: city || '',
+      canton: canton || null,
+      postalCode: postalCode || '',
+      phone: '',
+      address: '',
+      maxTravelKm,
+      isActive,
+      profileTier: REPAIRER_PROFILE_TIER.COMMUNITY,
+      status: REPAIRER_STATUS.ACTIVE,
+    })
+    .onConflictDoUpdate({
+      target: repairerProfiles.userId,
+      set: {
+        description: bio || null,
+        hourlyRateCents,
+        acceptsGratis,
+        acceptsKulturlegi,
+        serviceDeliveryTypes: serviceTypes.length > 0 ? serviceTypes : null,
+        city: city || '',
+        canton: canton || null,
+        postalCode: postalCode || '',
+        maxTravelKm,
+        isActive,
+        // Never demote a verified professional via this self-service endpoint.
+        profileTier: sql`CASE WHEN ${repairerProfiles.profileTier} = ${REPAIRER_PROFILE_TIER.PROFESSIONAL} THEN ${REPAIRER_PROFILE_TIER.PROFESSIONAL} ELSE ${REPAIRER_PROFILE_TIER.COMMUNITY} END`,
+        updatedAt: sql`NOW()`,
+      },
+    })
+
+  // Replace skills: delete existing, insert new.
+  await db.delete(userSkills).where(eq(userSkills.userId, userId))
+  if (skills.length > 0) {
+    await db.insert(userSkills).values(
+      skills.map((skillId) => ({
+        userId,
+        skillId,
+        categoryId: getCategoryForSkill(skillId),
+      })),
+    )
+  }
 }
